Refactor install into command for programatic use in testing
