+++ src/commands/toggleLanguageServer.ts
@@ -3,7 +3,7 @@ import { IndexAdapter } from "../index/index-adapter";
 import { Command, CommandType } from "./command";
 import { getConfiguration } from "../configuration";
 import { ExperimentalLanguageClient } from "../languageclient";
-import _ = require("lodash");
+import * as _ from "lodash";
 
 export class ToggleLanguageServerCommand extends Command {
   public static readonly CommandName = "toggleLanguageServer";
